+++ GridKit/Model/PhasorDynamics/SystemModelDataJSONParser.hpp
@@ -148,6 +148,12 @@ namespace GridKit
           raw_component.get_to(reeca);
           sm.reeca.push_back(reeca);
         }
+        else if (kind == "Reecb")
+        {
+          typename SystemModelData<RealT, IdxT>::ReecbDataT reecb;
+          raw_component.get_to(reecb);
+          sm.reecb.push_back(reecb);
+        }
         else if (kind == "Tgov1")
         {
           typename SystemModelData<RealT, IdxT>::Tgov1DataT gov;